Rewrite stuff to serializer factory
